+++ includes/Core/Loader.php
@@ -28,6 +28,8 @@ class Loader {
         \WeblockWidgets\Widgets\Social\FacebookFeed::instance()->init();
         \WeblockWidgets\Widgets\Social\YoutubeGallery::instance()->init();
         \WeblockWidgets\Widgets\Social\TikTokFeed::instance()->init();
+        \WeblockWidgets\Widgets\Social\TwitterFeed::instance()->init();
+        \WeblockWidgets\Widgets\Social\PinterestFeed::instance()->init();
         \WeblockWidgets\Widgets\Tools\GoogleMaps::instance()->init();
 
         add_action( 'wp_enqueue_scripts', [ $this, 'enqueue_assets' ] );